+++ check_workflow/cli.py
@@ -1,10 +1,11 @@
 import argparse
+from pathlib import Path
 
 from check_workflow.gh_api import fetch_workflows
-from check_workflow.workflow import format_outdated, report_outdated
+from check_workflow.workflow import fetch_local, format_outdated, report_outdated
 
 
-def _report_pipeline(org: str, repo: str, root: str, branch: str, markdown: bool) -> None:
+def _remote_report_pipeline(org: str, repo: str, root: str, branch: str, markdown: bool) -> None:
     workflows = fetch_workflows(owner=org, repo_name=repo, workflow_root=root, branch=branch)
     outdated = report_outdated(workflows)
 
@@ -12,26 +13,56 @@ def _report_pipeline(org: str, repo: str, root: str, branch: str, markdown: bool
         print(format_outdated(outdated, markdown=markdown))
 
 
+def _local_report_pipeline(root: Path, markdown: bool) -> None:
+    workflows = fetch_local(root)
+    outdated = report_outdated(workflows)
+
+    if outdated:
+        print(format_outdated(outdated, markdown=markdown))
+
+
 def main() -> None:  # noqa: D103
-    parser = argparse.ArgumentParser(
-        "CheckWorkflow", formatter_class=argparse.ArgumentDefaultsHelpFormatter
+    parser = argparse.ArgumentParser("CheckWorkflow")
+    subparsers = parser.add_subparsers(dest="subcommand")
+
+    # Query local project
+    local_sub = subparsers.add_parser(
+        "local", help="Query local project", formatter_class=argparse.ArgumentDefaultsHelpFormatter
     )
-    parser.add_argument("org", type=str, help="Query repository parent")
-    parser.add_argument("repo", type=str, help="Query repository")
-    parser.add_argument(
+    local_sub.add_argument(
+        "-r", "--root", type=Path, default="./.github/workflows/", help="Workflow root"
+    )
+    local_sub.add_argument(
+        "-m", "--markdown", action="store_true", help="Format report as markdown"
+    )
+
+    # Query remote repo
+    remote_sub = subparsers.add_parser(
+        "remote",
+        help="Query remote repository",
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
+    remote_sub.add_argument("org", type=str, help="Query repository parent")
+    remote_sub.add_argument("repo", type=str, help="Query repository")
+    remote_sub.add_argument("-b", "--branch", type=str, default="main", help="Query branch")
+    remote_sub.add_argument(
         "-r", "--root", type=str, default=".github/workflows/", help="Workflow root"
     )
-    parser.add_argument("-b", "--branch", type=str, default="main", help="Query branch")
-    parser.add_argument("-m", "--markdown", action="store_true", help="Format report as markdown")
+    remote_sub.add_argument(
+        "-m", "--markdown", action="store_true", help="Format report as markdown"
+    )
 
     args = parser.parse_args()
-    _report_pipeline(
-        org=args.org,
-        repo=args.repo,
-        root=args.root,
-        branch=args.branch,
-        markdown=args.markdown,
-    )
+    if args.subcommand == "local":
+        _local_report_pipeline(root=args.root, markdown=args.markdown)
+    else:
+        _remote_report_pipeline(
+            org=args.org,
+            repo=args.repo,
+            root=args.root,
+            branch=args.branch,
+            markdown=args.markdown,
+        )
 
 
 if __name__ == "__main__":