+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
+## `[v1.1.0]`
+
+### Changed
+
+* #5 CLI execution has been split into `local` and `remote` subcommands; `CheckWorkflow remote <...>` now provides the previous functionality
+
+### Added
+
+* #5 Add parsing of local project workflows using `CheckWorkflow local <...>`
+
 ## `[v1.0.0]`
 
 Initial release 🎉